# sql: don't construct empty SessionData.CustomOptions