@nvanbenschoten 2022-01-10
It does not look like this needs a backport. The offending code does not exist on release-21.2.